app/requirements/base.txt
--trusted-host 10.16.0.230
--extra-index-url http://10.16.0.230:8080/pypi/simple

Django==4.2
djangorestframework==3.14.0
django-cors-headers==4.2.0
django-health-check==3.17.0
django-session-timeout==0.1.0
django-crispy-forms==2.0
django-webpack-loader==2.0.1
crispy-bootstrap5==0.7
psycopg2-binary==2.9.6
python-decouple==3.4
requests==2.28.1
whitenoise==6.4.0
ntt-portal-library==6.0.49
datatable-ajax-request-parser==1.0.5
kombu==5.3.3
